tutor_group_system/requirements/developer.txt
# Install the main set of requirements
-r ./base.txt

# Additional requirements useful for developers of tutor group system

# Django debugging
django-debug-toolbar
